+++ CHANGELOG.md
@@ -3,9 +3,8 @@
 ## Version 0.4.2-dev (changes since 0.4.2 go here)
 
 ### Features
-* Added `skbio.sequence.replace` for replacing items in a `Sequence` with a given character. ([#1222](https://github.com/biocore/scikit-bio/issues/1222))
-=======
 * Added `skbio.sequence.distance.kmer_distance` for computing the kmer distance between two sequences. ([#913](https://github.com/biocore/scikit-bio/issues/913))
+* Added `skbio.sequence.replace` for assigning a character to positions in a `Sequence`. ([#1222](https://github.com/biocore/scikit-bio/issues/1222))
 
 ### Backward-incompatible changes [stable]